fix sync decoder conditional compilation things
